move config validator ts file
