+++ projects/core/src/config/config-validator/config-validator.ts
