+++ projects/core/src/config/index.ts
@@ -3,5 +3,5 @@ export {
   TestConfigModule,
   TestConfigModuleOptions,
 } from './test-config.module';
-export * from './utils/config-validator';
+export * from './config-validator/config-validator';
 export * from './config-initializer/index';
