+++ projects/core/src/occ/occ.module.ts
@@ -1,6 +1,6 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { Config, provideConfig } from '../config/config.module';
-import { provideConfigValidator } from '../config/utils/config-validator';
+import { provideConfigValidator } from '../config/config-validator/config-validator';
 import { AsmOccModule } from './adapters/asm/asm-occ.module';
 import { CartOccModule } from './adapters/cart/cart-occ.module';
 import { CheckoutOccModule } from './adapters/checkout/checkout-occ.module';
